MAINT: Rename to q2studio
