--- .babelrc
@@ -1,9 +1,0 @@
-{
-  "presets": [
-    "es2015",
-    "react"
-  ],
-  "plugins": [
-    "transform-object-rest-spread"
-  ]
-}
