+++ README.md
@@ -1,5 +1,5 @@
-# qiime-studio
-A web interface for QIIME 2. 
+# q2studio
+A web interface for QIIME 2.
 
 For user install instructions please visit [Installing and using QIIME 2](http://2.qiime.org/docs/interfaces/qiime-studio.html)
