+++ app/index.html
@@ -12,7 +12,7 @@ The full license is in the file LICENSE, distributed with this software.
 <html>
   <head>
     <meta charset="utf-8">
-    <title>QIIME Studio</title>
+    <title>QIIME 2 Studio</title>
   </head>
   <body>
     <div id="root"></div>
