+++ app/main.js
@@ -37,7 +37,7 @@ const startRestAPI = (callback) => {
                         .find(binaryPath => binaryPath.search('conda') !== -1);
     const condaBin = path.dirname(python);
 
-    const api = spawn('python', ['-u', '-m', 'qiime_studio'], {
+    const api = spawn('python', ['-u', '-m', 'q2studio'], {
         env: {
             ...process.env,
             // prepend conda bin to PATH
