+++ q2studio/__main__.py
@@ -6,9 +6,9 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
-from qiime_studio import start_server
+from q2studio import start_server
 
-# Allows `python -m qiime_studio` to start the server. This avoids polluting
+# Allows `python -m q2studio` to start the server. This avoids polluting
 # the user's namespace with "useless" binary.
 if __name__ == '__main__':
     try:
